♻️ refactor: 优化 download option ui
