+++ src/utils/common/map.py
@@ -117,6 +117,8 @@ status_code_map = {
     610: "执行外部命令失败",
     611: "下载失败",
     612: "下载失败达到最大重试次数",
+    613: "用户终止解析操作",
+    614: "该内容采用 WideVine DRM 技术加密，不支持下载",
     62002: "稿件不可见",
     62004: "稿件审核中",
     62012: "仅 UP 主自己可见",
