+++ src/utils/parse/episode/episode_v2.py
@@ -102,6 +102,7 @@ class EpisodeInfo:
             "collection_title": episode.get("collection_title", ""),
             "series_title": episode.get("series_title", ""),
             "interact_title": episode.get("interact_title", ""),
+            "parent_title": episode.get("parent_title", ""),
             "area": episode.get("area", ""),
             "zone": episode.get("zone", ""),
             "subzone": episode.get("subzone", ""),
