+++ src/utils/parse/episode/favlist.py
@@ -1,20 +1,27 @@
+from utils.config import Config
+from utils.common.enums import EpisodeDisplayType
+
 from utils.parse.episode.episode_v2 import EpisodeInfo, Filter
 from utils.parse.episode.video import Video
 from utils.parse.episode.bangumi import Bangumi
 
 class FavList:
     target_bvid: str = ""
+    parent_title: str = ""
     season_dict: dict = {}
 
     @classmethod
-    def parse_episodes(cls, info_json: dict, season_dict: dict):
+    def parse_episodes(cls, info_json: dict, season_dict: dict, parent_title: str):
         cls.season_dict = season_dict
+        cls.parent_title = parent_title
 
         EpisodeInfo.clear_episode_data()
 
+        Config.Misc.episode_display_mode = EpisodeDisplayType.All.value
+
         cls.favlist_parser(info_json)
 
-        Filter.episode_display_mode(reset = True)
+        Filter.episode_display_mode()
 
     @classmethod
     def favlist_parser(cls, info_json: dict):
@@ -33,17 +40,14 @@ class FavList:
         bvid = episode_info.get("bvid")
         video_info = cls.season_dict["video"][bvid]
 
-        if "ugc_season" in video_info:
-            Video.ugc_season_parser(video_info.copy())
-        else:
-            Video.pages_parser(video_info.copy())
+        Video.parse_episodes(video_info, cls.parent_title)
 
     @classmethod
     def bangumi_parser(cls, episode_info: dict):
         season_id = episode_info["ogv"]["season_id"]
         info_json = cls.season_dict["bangumi"][season_id]
 
-        Bangumi.parse_episodes(info_json)
+        Bangumi.parse_episodes(info_json, parent_title = cls.parent_title)
 
     @classmethod
     def get_entry_info(cls, episode: dict):
